Rename AdminDashboard.js/scss files to index. Move them to the dashboard directory inside the admin component.
