+++ client/src/components/admin/dashboard/index.scss
